refactor: 서비스 로직 추상화
