+++ src/css/index.css
@@ -44,7 +44,6 @@ body {
 .remains {
   border: 1px dashed #00bcd4;
   border-radius: 0.2rem;
-  background-color: #cdffff6b;
   text-align: center;
 }
