+++ src/js/handlers/input/inputNumberDuplicate.js
@@ -1,0 +1,30 @@
+import { hasDuplicate } from '../../lib/utils/validation.js';
+import { DUPLICATE_INPUT_NUMBER } from '../../lib/constants/alertMessage.js';
+
+const inputNumberDuplicateHandler = ({ target }) => {
+  const sliceRange = { begin: 0, end: 6 };
+  let inputArray = [...target.parentElement.children];
+
+  if (target.parentElement.children.length !== 7) {
+    sliceRange.begin = 0;
+    sliceRange.end = 7;
+    inputArray = [...target.form];
+  }
+
+  const isUniqueNumber = inputArray
+    .slice(sliceRange.begin, sliceRange.end)
+    .map(child => {
+      if (child.value !== '') {
+        return child.value;
+      }
+    })
+    .filter(child => child !== undefined);
+
+  if (hasDuplicate(isUniqueNumber)) {
+    target.value = '';
+    target.focus();
+    alert(DUPLICATE_INPUT_NUMBER);
+  }
+};
+
+export default inputNumberDuplicateHandler;
